Move to /editForward
