+++ lib/mayaUsd/CMakeLists.txt
@@ -321,6 +321,9 @@ install(FILES ${CMAKE_CURRENT_BINARY_DIR}/lib/usd/plugInfo.json
 # -----------------------------------------------------------------------------
 add_subdirectory(base)
 add_subdirectory(commands)
+if(AdskUsdEditForward_FOUND)
+    add_subdirectory(editForward)
+endif()
 add_subdirectory(fileio)
 add_subdirectory(listeners)
 add_subdirectory(nodes)
