+++ lib/mayaUsd/editForward/CMakeLists.txt
@@ -1,0 +1,25 @@
+# -----------------------------------------------------------------------------
+# sources
+# -----------------------------------------------------------------------------
+target_sources(${PROJECT_NAME}
+    PRIVATE
+        MayaUsdEditForwardCommand.cpp
+        MayaUsdEditForwardHost.cpp
+)
+
+set(HEADERS
+    MayaUsdEditForwardCommand.h
+    MayaUsdEditForwardHost.h
+)
+
+# -----------------------------------------------------------------------------
+# promoted headers
+# -----------------------------------------------------------------------------
+mayaUsd_promoteHeaderList(HEADERS ${HEADERS} SUBDIR editForward)
+
+# -----------------------------------------------------------------------------
+# install
+# -----------------------------------------------------------------------------
+install(FILES ${HEADERS}
+    DESTINATION ${CMAKE_INSTALL_PREFIX}/include/${PROJECT_NAME}/editForward
+)
