+++ lib/mayaUsd/editForward/MayaUsdEditForwardCommand.cpp
@@ -35,7 +35,7 @@ MayaUsdEditForwardCommand::create(const std::function<void()>& callback)
     return create(Callbacks { callback });
 }
 
-std::string MayaUsdEditForwardCommand::commandString() const { return "Forward USD Edits"; }
+std::string MayaUsdEditForwardCommand::commandString() const { return "ForwardUSDEdits"; }
 
 void MayaUsdEditForwardCommand::undo() { _undoableItem.undo(); }
 void MayaUsdEditForwardCommand::redo() { _undoableItem.redo(); }
