+++ lib/mayaUsd/editForward/MayaUsdEditForwardCommand.h
